requirements.txt
accelerate==0.33.0
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
Automat==22.10.0
beartype==0.16.4
blessed==1.20.0
buildtools==1.0.6
bitsandbytes
# causal-conv1d # ==1.1.3.post1
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
clean-fid==0.1.35
click==8.1.7
clip # ==1.0
cmake==3.28.1
colorama==0.4.6
coloredlogs==15.0.1
constantly==23.10.4
contourpy # ==1.2.0
crcmod==1.7
cryptography==41.0.7
cycler==0.12.1
decorator==5.1.1
decord==0.6.0
diffusers==0.26.3
docopt==0.6.2
easydict==1.11
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
fairscale==0.4.13
fastapi==0.109.0
# ffmpeg==1.4
# ffmpy==0.3.1
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
ftfy==6.1.3
furl==2.1.3
gpustat==1.1.1
gradio==4.14.0
gradio_client==0.8.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.2
humanfriendly==10.0
hyperlink==21.0.0
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib-metadata==7.0.1
importlib-resources==6.1.1
incremental==22.10.0
ipdb # ==0.13.13
ipython # ==8.18.1
jedi==0.19.1
Jinja2==3.1.3
jmespath==0.10.0
joblib==1.3.2
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kornia==0.7.1
lazy_loader==0.3
lightning-utilities==0.10.0
lit==17.0.6
lpips==0.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib # ==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
# motion-vector-extractor # ==1.0.6
mpmath==1.3.0


mlflow
multidict==6.0.4
mypy-extensions==1.0.0
networkx # ==3.2.1
ninja==1.11.1.1
numpy # ==1.26.3
# nvidia-cublas-cu11==11.10.3.66
# nvidia-cublas-cu12==12.1.3.1
# nvidia-cuda-cupti-cu11==11.7.101
# nvidia-cuda-cupti-cu12==12.1.105
# nvidia-cuda-nvrtc-cu11==11.7.99
# nvidia-cuda-nvrtc-cu12==12.1.105
# nvidia-cuda-runtime-cu11==11.7.99
# nvidia-cuda-runtime-cu12==12.1.105
# nvidia-cudnn-cu11==8.5.0.96
# nvidia-cudnn-cu12==8.9.2.26
# nvidia-cufft-cu11==10.9.0.58
# nvidia-cufft-cu12==11.0.2.54
# nvidia-curand-cu11==10.2.10.91
# nvidia-curand-cu12==10.3.2.106
# nvidia-cusolver-cu11==11.4.0.1
# nvidia-cusolver-cu12==11.4.5.107
# nvidia-cusparse-cu11==11.7.4.91
# nvidia-cusparse-cu12==12.1.0.106
# nvidia-ml-py==12.535.133
# nvidia-nccl-cu11==2.14.3
# nvidia-nccl-cu12==2.19.3
# nvidia-nvjitlink-cu12==12.3.101
# nvidia-nvtx-cu11==11.7.91
# nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnxruntime
open-clip-torch==2.24.0
# opencv-python==4.5.3.56
# opencv-python-headless==4.9.0.80
orderedmultidict==1.0.1
orjson==3.9.10
oss2==2.18.4
packaging # ==23.2
# pandas==2.1.4
parso==0.8.3
pexpect==4.9.0
pillow==10.2.
piq==0.8.0
pkgconfig==1.5.5
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
ptflops==0.7.2.2
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
Pygments==2.17.2
pynvml==11.5.0
pyparsing==3.1.1
pyre-extensions==0.0.29
python-dateutil==2.8.2
python-multipart==0.0.6
pytorch-lightning==2.1.3
pytz==2023.3.post1
PyYAML==6.0.1
redo==2.0.4
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rotary-embedding-torch==0.5.3
rpds-py==0.17.1
ruff==0.2.0
safetensors==0.4.1
scikit-image # ==0.22.0
scikit-learn # ==1.4.0
scipy # ==1.11.4
semantic-version==2.10.0
sentencepiece==0.1.99
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
sk-video==1.1.10
sniffio==1.3.0
SQLAlchemy==2.0.27
stack-data==0.6.3
starlette==0.35.1
sympy==1.12
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile # ==2023.12.9
timm==0.9.12
tokenizers==0.15.0
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
# torch==2.0.1+cu118
# torchaudio==2.0.2+cu118
# torchdiffeq==0.2.3
# torchmetrics==1.3.0.post0
# torchsde==0.2.6
# torchvision==0.15.2+cu118
tqdm==4.66.1
traitlets==5.14.1
trampoline==0.1.2
transformers==4.36.2
triton==2.0.0
Twisted==23.10.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.26.0
wcwidth==0.2.13
websockets==11.0.3
xformers==0.0.20
yarl==1.9.4
zipp==3.17.0
zope.interface==6.2
onnxruntime-gpu==1.13.1


# pip install simplejson rotary_embedding_torch fairscale imageio-ffmpeg accelerate==0.33.0 diffusers==0.26.3 onnxruntime-gpu